04/02/2019

+++ src/co/com/samtel/cargue/enumeraciones/tables/TypeBigCalificacionCarteraClienteColumn.java
@@ -12,7 +12,7 @@ public enum TypeBigCalificacionCarteraClienteColumn implements IColumn {
 	FUERZA_COMERCIAL(3, "I_fuerza_comercial", String.class),
 	OFICINA(4, "I_oficina", String.class),
 	DESCRIPCION_LINEA(5, "S_descripcion_linea", String.class),
-	COD_OPERACION(6, "I_cod_operacion", Long.class),
+	COD_OPERACION(6, "I_cod_operacion", Double.class),
 	CALIFICACION_ANTERIOR(7, "S_calificacion_anterior", String.class),
 	CALIFICACION_ACTUAL(8, "S_calificacion_actual", String.class),
 	FECHA_INICIO_CRE(9, "D_fecha_inicio_cre", Date.class),
